+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
  
 ## [v0.6.0] - 2021-XX-XX
 
-* Add caseinst GetList
+* Add caseinst module
 
 ## [v0.5.0] - 2021-07-31